created index.html and script.js
Created a html skeleton for the to-do-list application, using "utf-8" value for the character encoding. A jQuery plugin and a script.js file is scripted onto the index.html, respectively-  for manipulating and traversing the DOM.

on index.html, added a <form> tag containing an input element. the input's value will be appended to a newly created list node, in the DOM.

 a <div> element is manipulated into a submit button, on index.html, with the onclick attribute listening for a 'click' event; calling the newItem() function created in script.js using jQuery.

selecting the li node, in script.js, and appending the input's value using jQuery, it is then appended onto the <ol> node (found on the index.html).

on script.js, a function is created for adding/ or removing the '.strike' class of an element. This will be referenced in an event handler.

after selecting and manipulating the <crossOutButton> node, it is appended to the variable li -which is selecting the <li> node that was created earlier on the DOM.

another function is created for referencing a 'click' event handler. This will add the '.delete' class upon a 'click' event on the <li> node.

lastly on script.js, for manipulating the <ol> node, we added the .sortable() method- allowing a change in its order.